+++ pypsa/components/transform.py
@@ -97,6 +97,7 @@ class ComponentsTransformMixin:
         The example is shown for Generator component, but the same applies to all
         component types.
 
+        >>> import pypsa
         >>> n = pypsa.Network()
         >>> c = n.components.generators
         >>> c